Account for upstream changes to ReSharper CLT.
The URL at which the ReSharper Command Line Tools are hosted was
changed upstream. Now a single ZIP archive is published containing the
tools for all operating systems.

Introduce a version number to invalidate the cache key for the
decompressed CLT. Doing so fixes a bug where the cache of the
decompressed CLT would previously be erroneously shared across operating
systems. Cross-platform sharing is now safe, but only for caches of the
new consolidated ZIP archives.

The main drawback of this change is that the ReSharper Command Line
Tools for all operating systems are downloaded, but it is unclear
whether or where OS-specific archives might be hosted.